+++ src/css/popup.css
@@ -3,9 +3,11 @@ body {
   height: 600px;
   width: 400px;
   position: relative;
-  /* margin: 0px auto; */
   font-family: arial;
   background-color: #f1f2f8;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
 }
 .center {
   text-align: center;
@@ -14,10 +16,9 @@ body {
 header {
   background-color: #192234;
   margin-left: -8px;
-  margin-top: -18px;
+  margin-top: -20px;
   height: 75px;
-  margin-bottom: 60px;
-  /* shadow */
+  margin-bottom: 25px;
   -webkit-box-shadow: 0 4px 6px -6px #222;
   -moz-box-shadow: 0 4px 6px -6px #222;
   box-shadow: 0 4px 6px -6px #222;
@@ -38,22 +39,110 @@ h1 {
   cursor: default;
 }
 
-.btn-upload {
-  width: 250px;
-  padding: 10px;
-  border: 2px solid #326d6e;
-  border-radius: 8px;
-  background-color: #449394;
-  font-size: large;
-  color: white;
-  text-align: center;
-  cursor: pointer;
+.pages {
+  margin-bottom: 25px;
 }
-.btn-upload:hover {
-  background-color: lightseagreen;
+.file {
+  font-family: "Open Sans", sans-serif;
+  font-size: 15px;
+  letter-spacing: 2px;
+  text-decoration: none;
+  text-transform: uppercase;
+  color: #000;
+  background-color: white;
+  width: 110px;
+  cursor: pointer;
+  border: 3px solid;
+  padding: 0.35em 1.2em;
+  box-shadow: 1px 1px 0px 0px, 2px 2px 0px 0px, 3px 3px 0px 0px, 4px 4px 0px 0px,
+    5px 5px 0px 0px;
+  position: relative;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+}
+.file:active {
+  box-shadow: 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px;
+  top: 2px;
+  left: 5px;
+}
+@media (min-width: 768px) {
+  .file {
+    padding: 0.25em 0.75em;
+  }
+}
+.folder {
+  display: inline;
+  font-family: "Open Sans", sans-serif;
+  font-size: 15px;
+  margin-left: 40px;
+  letter-spacing: 2px;
+  text-decoration: none;
+  text-transform: uppercase;
+  color: #000;
+  background-color: white;
+  width: 110px;
+  cursor: pointer;
+  border: 3px solid;
+  padding: 0.35em 0.5em;
+  box-shadow: 1px 1px 0px 0px, 2px 2px 0px 0px, 3px 3px 0px 0px, 4px 4px 0px 0px,
+    5px 5px 0px 0px;
+  position: relative;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+}
+.folder:active {
+  box-shadow: 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px;
+  top: 2px;
+  left: 5px;
+}
+@media (min-width: 768px) {
+  .folder {
+    padding: 0.25em 0.75em;
+  }
+}
+.active{
+  box-shadow: 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px, 1px 1px 0px 0px;
+  top: 2px;
+  left: 5px;
 }
 
-/* loader */
+hr {
+  margin-bottom: 30px;
+  margin-left: -8px;
+}
+
+.btn-upload {
+  font-family: "Open Sans", sans-serif;
+  font-size: 20px;
+  letter-spacing: 2px;
+  text-decoration: none;
+  text-transform: uppercase;
+  color: #326d6e;
+  background-color: lightseagreen;
+  width: 120px;
+  cursor: pointer;
+  border: 3px solid;
+  padding: 0.25em 0.5em;
+  box-shadow: 1px 1px 0px 0px, 2px 2px 0px 0px, 3px 3px 0px 0px, 4px 4px 0px 0px,
+    5px 5px 0px 0px;
+  position: relative;
+  user-select: none;
+  -webkit-user-select: none;
+  touch-action: manipulation;
+}
+.btn-upload:active {
+  box-shadow: 0px 0px 0px 0px;
+  top: 5px;
+  left: 5px;
+}
+@media (min-width: 768px) {
+  .btn-upload {
+    padding: 0.25em 0.75em;
+  }
+}
+
 .loader {
   border: 10px solid #e0dada;
   border-radius: 50%;
@@ -62,31 +151,39 @@ h1 {
   height: 70px;
   display: none;
   position: absolute;
-  top: 230px;
+  top: 275px;
   left: 160px;
   -webkit-animation: spin 2s linear infinite;
   animation: spin 1s linear infinite;
 }
-
-/* Safari */
 @-webkit-keyframes spin {
-  0% { -webkit-transform: rotate(0deg); }
-  100% { -webkit-transform: rotate(360deg); }
+  0% {
+    -webkit-transform: rotate(0deg);
+  }
+  100% {
+    -webkit-transform: rotate(360deg);
+  }
 }
-
 @keyframes spin {
-  0% { transform: rotate(0deg); }
-  100% { transform: rotate(360deg); }
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
 }
 
 #qrcode {
   margin-top: 40px;
-  margin-bottom: 10px;
+  margin-bottom: 5px;
 }
 
 #svg-download {
   margin-bottom: 0px;
   margin-left: 15px;
+  -webkit-box-shadow: 0 2px 6px -6px #222;
+  -moz-box-shadow: 0 2px 6px -6px #222;
+  box-shadow: 0 2px 6px -6px #222;
 }
 
 .output-label {
@@ -105,13 +202,20 @@ h1 {
   margin-right: 8px;
   margin-left: 4px;
   font-size: medium;
-  border: 2px solid #CCCDCD;
+  border: 2px solid #cccdcd;
   border-radius: 5px;
   text-align: center;
   align-items: center;
+  -webkit-box-shadow: 0 2px 6px -6px #222;
+  -moz-box-shadow: 0 2px 6px -6px #222;
+  box-shadow: 0 2px 6px -6px #222;
+  user-select:text;
+  -webkit-user-select: text;
+  touch-action: manipulation;
 }
-#cid:hover, #link:hover {
-  background-color: #FBFAFA;
+#cid:hover,
+#link:hover {
+  background-color: #fbfafa;
 }
 svg {
   display: inline-flex;
@@ -122,18 +226,20 @@ svg {
   margin-top: 20px;
   margin-bottom: -12px;
   margin-right: 14px;
-  border: 2px solid #CCCDCD;
-  color: #326D6E;
+  border: 2px solid #cccdcd;
+  color: #326d6e;
   border-radius: 5px;
   text-align: center;
   align-items: center;
+  -webkit-box-shadow: 0 2px 6px -6px #222;
+  -moz-box-shadow: 0 2px 6px -6px #222;
+  box-shadow: 0 2px 6px -6px #222;
 }
 svg:hover {
   color: #449394;
-  background-color: #FBFAFA;
+  background-color: #fbfafa;
   cursor: pointer;
 }
-
 
 /* hide scrollbar but allow scrolling */
 #cid,
