+++ src/popup.jsx
@@ -1,10 +1,9 @@
 import React from "react";
 import { render } from "react-dom";
+import "./css/popup.css";
 
 function Popup() {
-  return (
-    <div></div>
-  );
+  return <div></div>;
 }
 
 render(<Popup />, document.getElementById("root"));
